Move mdsal-trace into its place
This is code incoming from the controller, we need to integrate it
into mdsal structure. This patch performs the preliminary move out
of opendaylight/md-sal directory.

While this code is being built, it is not part of the feature
repository yet, so as not to interfere with controller's
odl-mdsal-trace feature.

JIRA: MDSAL-412
Change-Id: I04cf3dcda5d990f5b1e45a3b3ecf15978a0db958
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>